GRPC/client/requirements.txt
Flask
grpcio
google
protobuf